render weathercards
