+++ src/containers/Dashboard.js
@@ -1,14 +1,14 @@
-import { connect } from 'react-redux';
-import Dashboard from '../components/Dashboard';
-import * as actionCreators from '../actions/actionCreators';
-import { bindActionCreators } from 'redux';
-
-const mapStateToProps = (state) => {
- return state;
-};
-
-const mapDispatchToProps = (dispatch) => {
- return bindActionCreators(actionCreators, dispatch);
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Dashboard);
+// import { connect } from 'react-redux';
+// import Dashboard from '../components/Dashboard';
+// import * as actionCreators from '../actions/actionCreators';
+// import { bindActionCreators } from 'redux';
+//
+// const mapStateToProps = (state) => {
+//  return state;
+// };
+//
+// const mapDispatchToProps = (dispatch) => {
+//  return bindActionCreators(actionCreators, dispatch);
+// };
+//
+// export default connect(mapStateToProps, mapDispatchToProps)(Dashboard);
